+++ roguelike/Screen.h
@@ -1,5 +1,6 @@
 #ifndef __SCREEN_h
 #define __SCREEN_h
+#include <Arduino.h>
 
 class Screen
 {
@@ -11,8 +12,8 @@ class Screen
     }
     
     void Init();
-    void Draw(short xpos, short ypos, short width, short height, int* pixels);
-    void DrawPixel(short xpos, short ypos, int pixel);
+    void Draw(short xpos, short ypos, short width, short height, uint16_t* pixels);
+    void DrawPixel(short xpos, short ypos, uint16_t pixel);
     void DrawRect(short xpos, short ypos, short width, short height, int color);
     int Width();
     int Height();
